fix sqs_client
