+++ rust/cloud-storage/document_storage_service/src/main.rs
@@ -114,21 +114,6 @@ maybe_env_vars! {
     struct SnsApnsVoipPlatformArn;
 }
 
-const SQS_MAX_MESSAGES_PER_RECEIVE: i32 = 10;
-const SQS_MAX_WAIT_TIME_SECONDS: i32 = 20;
-
-fn validate_webhook_queue_configuration(config: &Config) -> anyhow::Result<()> {
-    anyhow::ensure!(
-        (1..=SQS_MAX_MESSAGES_PER_RECEIVE).contains(&config.webhook_queue_max_messages),
-        "WEBHOOK_QUEUE_MAX_MESSAGES must be between 1 and {SQS_MAX_MESSAGES_PER_RECEIVE}"
-    );
-    anyhow::ensure!(
-        (0..=SQS_MAX_WAIT_TIME_SECONDS).contains(&config.webhook_queue_wait_time_seconds),
-        "WEBHOOK_QUEUE_WAIT_TIME_SECONDS must be between 0 and {SQS_MAX_WAIT_TIME_SECONDS}"
-    );
-    Ok(())
-}
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     MacroEntrypoint::default().init();
@@ -146,7 +131,6 @@ async fn main() -> anyhow::Result<()> {
         .resolve_remote_secrets(env, &secretsmanager_client)
         .await
         .context("expected to be able to resolve config secrets")?;
-    validate_webhook_queue_configuration(&config)?;
 
     tracing::trace!("initialized config");
 
@@ -201,13 +185,15 @@ async fn main() -> anyhow::Result<()> {
     let document_delete_queue = macro_queues::DocumentDeleteQueue::new();
     let contacts_queue = macro_queues::ContactsQueue::new();
     let notification_queue = macro_queues::NotificationIngressQueue::new();
-    let webhook_event_queue = macro_queues::WebhookEventQueue::new();
+    let webhook_event_queue = webhook::outbound::SqsWebhookQueue::new(
+        aws_sdk_sqs::Client::new(&aws_config),
+        macro_queues::WebhookEventQueue::new().to_string(),
+        config.webhook_queue_max_messages,
+        config.webhook_queue_wait_time_seconds,
+    );
     let sqs_client = sqs_client::SQS::new(aws_sdk_sqs::Client::new(&aws_config))
         .search_event_queue(&search_event_queue)
-        .document_delete_queue(&document_delete_queue)
-        .webhook_event_queue(&webhook_event_queue)
-        .webhook_event_queue_max_messages(config.webhook_queue_max_messages)
-        .webhook_event_queue_wait_time_seconds(config.webhook_queue_wait_time_seconds);
+        .document_delete_queue(&document_delete_queue);
 
     let contacts_ingress = Arc::new(contacts::domain::service::SqsContactsIngress {
         queue: contacts::outbound::ingress::SqsContactsQueue::new(
@@ -617,7 +603,7 @@ async fn main() -> anyhow::Result<()> {
         webhook::domain::ingestion::WebhookEventIngestionServiceImpl::new(
             entity_access_service.clone(),
             webhook_repository,
-            sqs_client.clone(),
+            webhook_event_queue.clone(),
         );
     let webhook_delivery_repository =
         webhook::outbound::PgWebhookDeliveryRepository::new(db.clone());
@@ -626,7 +612,7 @@ async fn main() -> anyhow::Result<()> {
         webhook_http_client,
     );
     let webhook_worker = webhook::inbound::worker::WebhookEventWorker::new(
-        sqs_client.clone(),
+        webhook_event_queue,
         webhook_delivery_service,
     );
     let webhook_worker_task = tokio::spawn(async move {
