+++ rust/cloud-storage/webhook/src/outbound.rs
@@ -4,8 +4,10 @@ pub mod http_delivery;
 pub mod http_validator;
 pub mod pg_delivery_repository;
 pub mod pg_repository;
+pub mod sqs_queue;
 
 pub use http_delivery::ReqwestWebhookDeliveryClient;
 pub use http_validator::ReqwestWebhookValidationClient;
 pub use pg_delivery_repository::PgWebhookDeliveryRepository;
 pub use pg_repository::PgRepository;
+pub use sqs_queue::SqsWebhookQueue;
